+++ src/pages/articles/[id].page.tsx
@@ -11,10 +11,10 @@ import Thumbnail from "@/components/atoms/Thumbnail";
 import { HtmlHeadBase, HtmlHeadJsonLd } from "@/components/functions/meta";
 import ArticleLayout from "@/components/layouts/ArticleLayout";
 import { TagListColored } from "@/components/molecules/TagList";
+import { getSafeDate } from "@/lib/date";
 import { mdx2html } from "@/lib/mdx";
 import { getPickupArticles, getPopularArticles, getRelatedArticles } from "@/services/article";
 import type { TArticle, TCategory, TConfig, TPickup, TRankedArticle } from "@/types";
-import { getSafeDate } from "@/utils/date";
 import { fetchArticle, fetchArticles, fetchCategories, fetchConfig } from "@/utils/fetcher";
 import { formatPageTitle, formatPageUrl, getExcerpt } from "@/utils/formatter";
 import { getBackLinks, urlTable } from "@/utils/paths/url";
